Do not indirectly encode JSON
To really get throughput out of serde we are incentivized to encode JSON arrays
as a single Vec. This commit makes that switch. To accomplish this we have to
estimate the encoded size but, done, allows us to use the vec batch.

I will follow this commit up with tests to bound the estimation a little more
closely.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>